feat: improved health chart tooltip
